Extract ChatComposer to fix composer keystroke re-renders
Typing in the composer caused the entire ChatView (~4900 lines) to
re-render on every keystroke because `useComposerThreadDraft` subscribed
to the full draft including `prompt`. This cascaded to MessagesTimeline,
ChatHeader, BranchToolbar, PlanSidebar, and terminal drawers.

- Extract `ChatComposer` (memo'd) that owns the prompt/images/terminal
  context store subscriptions, all composer-local state, and derived
  values (menu items, provider state, send state, debounced queries).
  ChatView now uses granular store selectors for only `runtimeMode`,
  `interactionMode`, and `activeProvider` — none of which change on
  keystroke.

- Extract `ExpandedImageDialog` (memo'd) for the image preview modal.

- Wrap `BranchToolbar` in `memo` and stabilize `onRevertUserMessage`
  with `useCallback` so already-memo'd children (MessagesTimeline,
  ChatHeader) properly skip re-renders.

- Expose `ChatComposerHandle` for cross-cutting operations (focus,
  cursor reset, terminal context insertion, send context).
