+++ apps/web/index.html
@@ -8,6 +8,7 @@
     <meta name="theme-color" content="#161616" />
     <link rel="icon" href="/favicon.ico" sizes="48x48" />
     <link rel="apple-touch-icon" href="/apple-touch-icon.png" />
+    <script crossorigin="anonymous" src="//unpkg.com/react-scan/dist/auto.global.js"></script>
     <script>
       (() => {
         const LIGHT_BACKGROUND = "#ffffff";
